PPU Scrolling
Implemented an incorrect & very inaccurate version of PPU scrolling.

I only learned after the fact that without rendering progressively on each scanline/pixel I can't accurately render games that modify data as the frame is rendered.

So committing this as the very naive approach and I'm going to revisit with a re-implementation later.