Major restructure to use Cartridge
Replaced ROM with Cartridge, this encompasses more of the Cartridge than just the ROM data. Mainly the different CPU/PPU reads which are delegated to a mapper (still only Mapper 0 implemented for now)

Along with this a significant restructure to the ownership of each component. Bus now also has less responsibility for mapping specific addresses instead it is mapping specific address spaces to the appropriate components.